move analyzers to separate files
