+++ tools/contract-analyzer/analyzers/reference_operator_analyzer.go
@@ -23,95 +23,8 @@ import (
 	"regexp"
 
 	"github.com/onflow/cadence/runtime/ast"
-	"github.com/onflow/cadence/runtime/sema"
 	"github.com/onflow/cadence/tools/analysis"
 )
-
-// Deprecated key functions analyzer
-
-var DeprecatedKeyFunctionsAnalyzer = (func() *analysis.Analyzer {
-
-	elementFilter := []ast.Element{
-		(*ast.InvocationExpression)(nil),
-	}
-
-	return &analysis.Analyzer{
-		Description: "Detects usages of the deprecated key management API. It will be removed in a future release.",
-		Requires: []*analysis.Analyzer{
-			analysis.InspectorAnalyzer,
-		},
-		Run: func(pass *analysis.Pass) interface{} {
-			inspector := pass.ResultOf[analysis.InspectorAnalyzer].(*ast.Inspector)
-
-			location := pass.Program.Location
-			elaboration := pass.Program.Elaboration
-			report := pass.Report
-
-			inspector.Preorder(
-				elementFilter,
-				func(element ast.Element) {
-					invocationExpression, ok := element.(*ast.InvocationExpression)
-					if !ok {
-						return
-					}
-
-					memberExpression, ok := invocationExpression.InvokedExpression.(*ast.MemberExpression)
-					if !ok {
-						return
-					}
-
-					memberInfo := elaboration.MemberExpressionMemberInfos[memberExpression]
-					member := memberInfo.Member
-					if member == nil {
-						return
-					}
-
-					if member.ContainerType != sema.AuthAccountType {
-						return
-					}
-
-					var replacement string
-					functionName := member.Identifier.Identifier
-					switch functionName {
-					case sema.AuthAccountAddPublicKeyField:
-						replacement = "keys.add"
-					case sema.AuthAccountRemovePublicKeyField:
-						replacement = "keys.revoke"
-					default:
-						return
-					}
-
-					report(
-						analysis.Diagnostic{
-							Location: location,
-							Range:    ast.NewRangeFromPositioned(nil, memberExpression.Identifier),
-							Category: "update recommended",
-							Message: fmt.Sprintf(
-								"deprecated function '%s' will get removed",
-								functionName,
-							),
-							SecondaryMessage: fmt.Sprintf(
-								"replace with '%s'",
-								replacement,
-							),
-						},
-					)
-				},
-			)
-
-			return nil
-		},
-	}
-})()
-
-func init() {
-	registerAnalyzer(
-		"deprecated-key-functions",
-		DeprecatedKeyFunctionsAnalyzer,
-	)
-}
-
-// Reference operator analyzer
 
 var ReferenceOperatorAnalyzer = (func() *analysis.Analyzer {
